requirements.txt
asgiref==3.5.2
autopep8==1.7.0
Django==4.1.1
django-extensions==3.2.1
djangorestframework==3.14.0
pycodestyle==2.9.1
python-decouple==3.6
pytz==2022.2.1
sqlparse==0.4.3
toml==0.10.2
tzdata==2022.4
